ultrasonics done - need spi added

+++ PIC24_Ultrasonic.X/frontUltrasonic.h
@@ -31,5 +31,5 @@ void __attribute__((__interrupt__, auto_psv)) _IC1Interrupt(void);
 void __attribute__((__interrupt__, auto_psv)) _IC2Interrupt(void);
 void __attribute__((__interrupt__, auto_psv)) _IC3Interrupt(void);
 
-unsigned convertToDistance(double time);
+double convertToDistance(double time);
 double findObject(void);
